Enable live thumbnail for all tabs
Without this, the thumbnail is refreshed only for the selected tab. It also causes that only the first thumbnail is visible when the application is launched. Setting the live-thumbnail property to True for all tabs solves the problem.

Reference: https://gnome.pages.gitlab.gnome.org/libadwaita/doc/main/class.TabOverview.html